#403 Better names for modules
