+++ twake/backend/node/src/core/platform/framework/index.ts
@@ -1,3 +1,5 @@
+import "reflect-metadata";
+
 export * from "./api";
 export * from "./decorators";
 export * from "./configuration";
