+++ twake/backend/node/src/services/websocket/provider.ts
@@ -1,5 +1,6 @@
 import socketIO from "socket.io";
+import { TwakeServiceProvider } from "../../core/platform/framework";
 
-export default interface WebSocketAPI {
+export default interface WebSocketAPI extends TwakeServiceProvider {
   getIo(): socketIO.Server;
 }
